+++ frontend/app/src/main/AndroidManifest.xml
@@ -19,9 +19,9 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".EditProfileActivity"
+            android:name=".MeetupActivity"
             android:exported="false"
-            android:label="@string/title_activity_edit_profile"
+            android:label="@MeetupActivity"
             android:theme="@style/Theme.Frontend" />
         <activity
             android:name=".MissionActivity"
