+++ frontend/app/src/main/java/com/example/frontend/model/PlaceModel.kt
@@ -1,0 +1,7 @@
+package com.example.frontend.model
+
+import com.google.android.gms.maps.model.LatLng
+
+data class PlaceModel (
+    var averagedLocation: LatLng?
+)
